:hammer: Split main.go into packages
