+++ README.md
@@ -151,7 +151,7 @@ Use [prebuilt binary](https://github.com/Rolix44/Kubestroyer/releases) or build 
 <!-- ROADMAP -->
 ## Roadmap
 
-- [ ] Split main into different packages
+- [x] Split main into different packages
 - [ ] Target selection
     - [ ] List of multiple targets
     - [ ] input file as target parameter
