+++ utils/display.go
@@ -1,0 +1,15 @@
+package utils
+
+var Toolname string = `
+ _   __      _               _                             
+| | / /     | |             | |                            
+| |/ / _   _| |__   ___  ___| |_ _ __ ___  _   _  ___ _ __ 
+|    \| | | | '_ \ / _ \/ __| __| '__/ _ \| | | |/ _ \ '__| v0.1
+| |\  \ |_| | |_) |  __/\__ \ |_| | | (_) | |_| |  __/ |   
+\_| \_/\__,_|_.__/ \___||___/\__|_|  \___/ \__, |\___|_|   
+                                            __/ |          
+                                           |___/ `
+var Author = `			                        By Rolix`
+var Split = `
+--------------------------------------------------------
+`
